pyproject.toml
[project]
name = "ids-rag"
version = "0.1.0"
description = "Local RAG system using Ollama and ChromaDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-chroma>=0.1.4",
    "chromadb>=0.5.5",
    "pyyaml>=6.0.2",
    "beautifulsoup4>=4.12.3",
    "pypdf>=5.0.0",
    "requests>=2.32.5",
]

[project.scripts]
ids-rag = "ids_rag.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ids_rag"]
